LPS-127321 Create constants with SCOPES of Deployment types
