+++ modules/apps/app-builder/app-builder-web/src/main/resources/META-INF/resources/js/pages/apps/constants.es.js
@@ -95,6 +95,23 @@ const FILTERS = [
 	},
 ];
 
+const SITE_MENU_SCOPE_VALUE = 'site_administration.content';
+
+const SCOPES = [
+	{
+		label: Liferay.Language.get('applications'),
+		value: ['applications_menu.applications'],
+	},
+	{
+		label: Liferay.Language.get('site-menu'),
+		value: [SITE_MENU_SCOPE_VALUE],
+	},
+	{
+		label: Liferay.Language.get('applications-and-site-menu'),
+		value: ['applications_menu.applications', SITE_MENU_SCOPE_VALUE],
+	},
+];
+
 export {
 	COLORS,
 	COLUMNS,
@@ -102,5 +119,7 @@ export {
 	DEPLOYMENT_TYPES,
 	FILTERS,
 	FILTER_NAMES,
+	SITE_MENU_SCOPE_VALUE,
 	STATUSES,
+	SCOPES,
 };
