+++ modules/apps/app-builder/app-builder-web/src/main/resources/META-INF/resources/js/pages/apps/edit/EditAppContext.es.js
@@ -27,7 +27,7 @@ export const UPDATE_NAME = 'UPDATE_NAME';
 export const UPDATE_SETTINGS_SCOPE = 'UPDATE_SETTINGS_SCOPE';
 export const UPDATE_WORKFLOW_PROCESS_ID = 'UPDATE_WORKFLOW_PROCESS_ID';
 
-const uppdateAppDeployment = (state, appDeploymentType, appDeployment) => ({
+const updateAppDeployment = (state, appDeploymentType, appDeployment) => ({
 	...state,
 	app: {
 		...state.app,
@@ -99,7 +99,7 @@ const reducer = (state, action) => {
 				},
 			};
 
-			return uppdateAppDeployment(state, PRODUCT_MENU, newAppDeployment);
+			return updateAppDeployment(state, PRODUCT_MENU, newAppDeployment);
 		}
 		case UPDATE_APP: {
 			return {
@@ -159,7 +159,7 @@ const reducer = (state, action) => {
 				},
 			};
 
-			return uppdateAppDeployment(state, PRODUCT_MENU, newAppDeployment);
+			return updateAppDeployment(state, PRODUCT_MENU, newAppDeployment);
 		}
 		case UPDATE_WORKFLOW_PROCESS_ID: {
 			return {
